chore: pull out getGitHubAppInstallationToken helper
